Initial Commit
This is the initial alpha release of LISA 0.1. It contains a Kernel suitable to represent and manipulate First Order Logic, Proofs in Sequent Calculus and mathematical Theories.
It also contains some development of proof tactics and some early theorems of set theory, a parser for TPTP problems and other less significant features.
This release also include a test suite for those elements.