Easy tactics (#62)
* New ProofSteps have intermediate representation without Premises and Bot, to allow a DSL where those are specified separately.
Compatible with the new Parser.
In existing proofs, all Sequent Calculus proof steps have been replaces by SC.ProofStep to disambiguate.

* Proofs now are being done imperatively; When a proofstep is created, it is added to a temporary list keeping track of the proof in the Library.
Arbitrary code can be executed in the middle of the proof, including showing the current state of the proof.
Further work include improving the tactics to make them need less parameters, an "auto" tactic, more tools.

* can use andThen for steps following the previous one without specifying the number
Mostly support imports: no need for using "using(...)" anymore. Instead, when calling "withImport" with a justification, it will both add the sequent as import and keep the justification for constructing the proof.
New higher level proof object. They maintain both the underlying SCProof and the current proofSteps and imports.
When a Proof step is (imperatively) added, it is checked. A (deduced) ProofStep can fail to return, if the parameters or situation it is in is not adequate.
Further work: Manage imports through subproofs, being able to refer to previous steps, having a uniform way to refer to previous steps and theory justifications.